hours from int to double

+++ src/main/java/com/guard/admin/database/entities/Schedule.java
@@ -25,7 +25,7 @@ public class Schedule {
 
     private Date startTime;
     private Date endTime;
-    private Integer hours;
+    private Double hours;
     private Double frequency;
     private String announces;
 }
